rename philosophers package to problem
